+++ public/js/src/module/gui.js
@@ -8,11 +8,13 @@ var support = require( 'enketo-core/src/js/support' );
 var settings = require( './settings' );
 var printForm = require( 'enketo-core/src/js/print' );
 var t = require( './translator' );
+var sniffer = require( './sniffer' );
 var dialog = require( './vex.dialog.custom' );
 var $ = require( 'jquery' );
 require( './plugin' );
 
 var pages;
+var homeScreenGuidance;
 var updateStatus;
 var feedbackBar;
 
@@ -82,6 +84,10 @@ function setEventHandlers() {
         var href = this.getAttribute( 'href' );
         return ( !href || href === '#' ) ? false : true;
     } );
+
+    if ( _getHomeScreenGuidance() ) {
+        $( '.form-header__button--homescreen' ).removeClass( 'hide' ).on( 'click', alertHomeScreenGuidance );
+    }
 
     $doc.on( 'xpatherror', function( ev, error ) {
         var email = settings[ 'supportEmail' ],
@@ -335,6 +341,41 @@ function alertLoadErrors( loadErrors, advice ) {
         t( 'alert.loaderror.msg2', params ) +
         '</p>' + errorStringHTML, t( 'alert.loaderror.heading', params )
     );
+}
+
+function alertHomeScreenGuidance() {
+    alert( _getHomeScreenGuidance(), t( 'alert.addtohomescreen.heading' ), 'normal' );
+}
+
+function _getHomeScreenGuidance() {
+    var imageClass1;
+    var imageClass2;
+    var guidanceKey;
+    var browser = sniffer.browser;
+    var os = sniffer.os;
+
+    if ( homeScreenGuidance ) {
+        // keep calm
+    } else if ( os.isIos() && browser.isSafari() ) {
+        imageClass1 = 'ios-safari';
+        homeScreenGuidance = t( 'alert.addtohomescreen.iossafari.msg', _getHomeScreenGuidanceObj( imageClass1 ) );
+    } else if ( os.isAndroid() && browser.isChrome() ) {
+        imageClass1 = 'android-chrome';
+        homeScreenGuidance = t( 'alert.addtohomescreen.androidchrome.msg', _getHomeScreenGuidanceObj( imageClass1 ) );
+    } else if ( os.isAndroid() && browser.isFirefox() ) {
+        imageClass1 = 'android-firefox-1';
+        imageClass2 = 'android-firefox-2';
+        homeScreenGuidance = t( 'alert.addtohomescreen.androidfirefox.msg', _getHomeScreenGuidanceObj( imageClass1, imageClass2 ) );
+    }
+
+    return homeScreenGuidance;
+}
+
+function _getHomeScreenGuidanceObj( imageClass1, imageClass2 ) {
+    return {
+        image1: ( imageClass1 ) ? '<span class="' + imageClass1 + '"/>' : '',
+        image2: ( imageClass2 ) ? '<span class="' + imageClass2 + '"/>' : ''
+    };
 }
 
 /**