+++ public/js/src/module/sniffer.js
@@ -1,24 +1,36 @@
 'use strict';
 
 var browser;
-
-/**
- * It is a sad state of affairs that we need this.
- */
+var os;
+var ua = navigator.userAgent;
+var platform = navigator.platform;
 
 browser = {
     isChrome: function() {
-        var matchedChrome = navigator.userAgent.match( /Chrome\/(\d+)/ );
-        var matchedEdge = navigator.userAgent.match( /Edge\// );
+        var matchedChrome = /chrome|crios\/(\d+)/i.test( ua );
+        var matchedEdge = /edge\//i.test( ua );
         // MS Edge pretends to be Chrome 42:
         // https://msdn.microsoft.com/en-us/library/hh869301%28v=vs.85%29.aspx
         return !matchedEdge && matchedChrome;
     },
-    isOnIos: function() {
-        return /iPad|iPhone|iPod/i.test( navigator.platform );
+    isSafari: function() {
+        return /^((?!chrome|android|fxios|crios|ucbrowser).)*safari/i.test( ua );
+    },
+    isFirefox: function() {
+        return /firefox|fxios/i.test( ua );
+    }
+};
+
+os = {
+    isIos: function() {
+        return /iPad|iPhone|iPod/i.test( ua );
+    },
+    isAndroid: function() {
+        return /android/i.test( ua );
     }
 };
 
 module.exports = {
-    browser: browser
+    browser: browser,
+    os: os
 };