+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ##### Added
 - Export functionality.
 
+##### Changed
+- Links are underlined.
+
 ##### Fixed
 - Under some conditions surveys/list and surveys/number return too many results.
 - Multiple span elements on same line get rendered as one span.