+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-[Unreleased] 
+[1.22.0] - 2016-01-01 
 ---------------------
 ##### Added
 - Export functionality.
+- Add-to-homescreen guidance for iOS/Safari, Android/Chrome and Android/Firefox.
 
 ##### Changed
 - Links are underlined.